Code cleanup from merges, carousel UI fix, exempt Facebook users form password complexity requirements
Facebook user’s complexity requirements is waived in a really hacky
way. It just saves a dummy string to pass the validation, then changes
their password to their Facebook key and saves it again, this time
checking the Facebook provider token to skip validation. This will be
fixed when the registrations process is fixed to force Facebook users
to register a password in the case of external account deletion.